+++ nwbwidgets/controllers/__init__.py
@@ -1,3 +1,5 @@
 from .time_window_controllers import StartAndDurationController, RangeController
 from .group_and_sort_controllers import GroupAndSortController
-from .misc import ProgressBar, make_trial_event_controller
+from .image_controllers import RotationController, ImShowController
+from .misc import ProgressBar, ViewTypeController, make_trial_event_controller
+from .multicontroller import MultiController
