+++ nwbwidgets/controllers/misc.py
@@ -11,15 +11,28 @@ class ProgressBar(tqdm_notebook):
         # self.container.children[0].layout = Layout(width="80%")
 
 
+class ViewTypeController(widgets.VBox):
+    controller_fields = ("view_type_toggle",)
+
+    def __init__(self):
+        super().__init__()
+
+        self.view_type_toggle = widgets.ToggleButtons(
+            options=[
+                ("Simplified", "Simplified"),
+                ("Detailed", "Detailed"),
+            ],  # Values set to strings for external readability
+        )
+        self.children = (self.view_type_toggle,)
+
+
 def make_trial_event_controller(trials, layout=None, multiple=False):
-    """Controller for which reference to use (e.g. start_time) when making time-aligned averages"""
+    """Controller for which reference to use (e.g. start_time) when making time-aligned averages."""
     trial_events = ["start_time"]
     if not np.all(np.isnan(trials["stop_time"].data)):
         trial_events.append("stop_time")
     trial_events += [
-        x.name
-        for x in trials.columns
-        if (("_time" in x.name) and (x.name not in ("start_time", "stop_time")))
+        x.name for x in trials.columns if (("_time" in x.name) and (x.name not in ("start_time", "stop_time")))
     ]
     kwargs = {}
     if layout is not None:
@@ -27,17 +40,10 @@ def make_trial_event_controller(trials, layout=None, multiple=False):
 
     if multiple:
         trial_event_controller = widgets.SelectMultiple(
-            options=trial_events,
-            value=["start_time"],
-            description='align to:',
-            disabled=False,
-            **kwargs
+            options=trial_events, value=["start_time"], description="align to:", disabled=False, **kwargs
         )
     else:
         trial_event_controller = widgets.Dropdown(
-            options=trial_events,
-            value="start_time",
-            description="align to: ",
-            **kwargs
+            options=trial_events, value="start_time", description="align to: ", **kwargs
         )
     return trial_event_controller
