+++ nwbwidgets/ophys/__init__.py
@@ -1,0 +1,3 @@
+from .plane_slice_visualization import PlaneSliceVisualization
+from .two_photon_series import TwoPhotonSeriesVizualization
+from .segmentation import PlaneSegmentation2DWidget, RoiResponseSeries
